requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
torch==2.1.0
torchaudio==2.1.0
transformers==4.35.0
librosa==0.10.1
soundfile==0.12.1
pydantic==2.5.0
python-multipart==0.0.6
numpy==1.24.3
python-dotenv==1.0.0
